Boost-only Options Processing
Until now options to zapps could be parsed from different sources and into different containers.
With these changes most of the options can now be entered on the command-line and are centrally stored in boost
program options. Commands initializing env-objects pass variable_maps to the env objects thus every option can be accessed via the optionValues attributes.
Path to configuration files can be added manually in the command-header file or supplied with the option config-file. Options given in the command line ovverride
options in the config files.