Removed "always-loading" of config file. Minor Bug Fixes
As discussed removed that the config file is always loaded. To load a shore.conf located in the folder of the executable just use --configuration-file or -cf. Can also be given a
concrete path. Minor Bug Fixes including a bug where when --trxs option was used the throughput print on console was printed to early.